requirements.txt
absl-py==2.3.1
affine==2.4.0
annotated-types==0.7.0
astunparse==1.6.3
cachetools==5.5.2
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
ftfy==6.3.1
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
hf-xet==1.1.7
huggingface-hub==0.34.4
humanfriendly==10.0
imageio==2.37.0
Jinja2==3.1.6
keras==2.15.0
kiwisolver==1.4.9
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
libclang==18.1.1
lpips==0.1.4
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml-dtypes==0.2.0
mpltern==1.0.4
mpmath==1.3.0
namex==0.1.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnxruntime==1.22.1
open_clip_torch==3.1.0
openai-clip==1.0.1
opencv-python==4.12.0.88
opensr_test==1.3.3
opt_einsum==3.4.0
optree==0.17.0
pandas==2.3.1
pillow==11.3.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydevd-pycharm==252.23892.439
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rasterio==1.4.3
regex==2025.7.34
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
safetensors==0.6.2
satalign==0.1.13
scikit-image==0.25.2
scipy==1.16.1
six==1.17.0
super-image==0.2.0
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tifffile==2025.6.11
timm==1.0.19
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.14.2
xarray==2025.7.1
